back/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
folium==0.15.0
geopy==2.4.1
python-dotenv==1.0.0
aiofiles==23.2.1
python-telegram-bot==20.7

